rename script varibale
